+++ resources/js/Components/Global/Uploaders/ImageUpload.vue
@@ -7,15 +7,25 @@
         <li>File Types accepted: <span class="text-orange-400">{{ fileTypes }}</span></li>
       </ul>
 
+      <div class="checkbox-container mb-4">
+        <input type="checkbox" id="removeExif" v-model="removeExif" class="checkbox checkbox-md">
+        <label for="removeExif" class="checkbox-label">Remove EXIF data</label>
+      </div>
+
+
       <file-pond
           name="image"
           ref="pond"
+          class-name="my-pond"
+          allow-multiple="false"
           label-idle="Click to choose file, or drag here..."
-          :server="serverOptions"
-          @init="filepondInitialized"
-          :accepted-file-types="fileTypes"
+          :server="`/upload?modelType=${props.modelType}&modelId=${props.modelId}&removeExif=${removeExif}`"
+          accepted-file-types="image/jpeg, image/png"
+          @init="handleFilePondInit"
           @processfile="handleProcessedFile"
           :max-file-size="maxSize"
+          allowRemove="false"
+          allowRevert="false"
       />
     </div>
   </div>
@@ -23,7 +33,6 @@
 </template>
 
 <script setup>
-import { useNotificationStore } from '@/Stores/NotificationStore'
 import vueFilePond from 'vue-filepond'
 import FilePondPluginFileValidateType from 'filepond-plugin-file-validate-type'
 import FilePondPluginFileValidateSize from 'filepond-plugin-file-validate-size'
@@ -32,18 +41,18 @@ import 'filepond/dist/filepond.min.css'
 import 'filepond-plugin-image-preview/dist/filepond-plugin-image-preview.min.css'
 import { ref } from 'vue'
 
-const notificationStore = useNotificationStore()
-
 let props = defineProps({
   image: Object,
   name: String,
   metadataKey: String,
   metadataValue: String,
   modelType: String,
-  modelId: String,
+  modelId: Number,
   maxSize: String,
   fileTypes: String,
 })
+
+const myFiles = ref(['']);
 
 const emit = defineEmits(['reloadImage', 'imageUploaded'])
 
@@ -54,26 +63,30 @@ const FilePond = vueFilePond(
 )
 
 // Define reactive variables
-const myFiles = ref([{ source: 'cat.jpeg', options: { type: 'local' } }]);
+const removeExif = ref(false);
 const maxSize = ref('30MB');
 const fileTypes = ref(['image/png', 'image/jpeg']);
-const modelType = 'yourModelType'; // Replace with your actual model type
-const modelId = 'yourModelId'; // Replace with your actual model ID
+// const modelType = 'yourModelType'; // Replace with your actual model type
+// const modelId = 'yourModelId'; // Replace with your actual model ID
 
 const serverOptions = ref({
   process: {
     url: `/upload?modelType=${props.modelType}&modelId=${props.modelId}`,
     method: 'POST',
     withCredentials: false,
-    headers: {},
-    onload: (response) => response.key,
-    onerror: (response) => response.data,
+    // headers: {},
+    // onload: (response) => response.key,
+    // onerror: (response) => response.data,
   },
 });
 
 
-function filepondInitialized() {
-  console.log('Filepond is ready!')
+// function filepondInitialized() {
+//   console.log('Filepond is ready!')
+// }
+
+function handleFilePondInit() {
+  console.log('FilePond has initialized');
 }
 
 function handleProcessedFile(error, file) {
@@ -89,3 +102,14 @@ function handleProcessedFile(error, file) {
 
 
 </script>
+
+<style scoped>
+.checkbox-container {
+  display: flex;
+  align-items: center; /* Align items vertically */
+}
+
+.checkbox {
+  margin-right: 8px; /* Space between the checkbox and the label */
+}
+</style>